add partner route
