+++ src/app.js
@@ -2,7 +2,7 @@ import express from 'express'
 import cors from 'cors'
 import bodyParser from 'body-parser'
 import morgan from 'morgan'
-import userRoute from './catalog/user/user.route'
+import partnerRoute from './catalog/partner/partner.route'
 
 const app = express()
 app.use(morgan('[:date[clf]] :method :url :status :response-time ms - :res[content-length]'))
@@ -18,7 +18,7 @@ app.use(cors({
   credentials: true //Turn on cookie HTTP through CORS
 }))
 
-app.use('/user', userRoute)
+app.use('/partner', partnerRoute)
 
 //handling 'Not found' error
 //if the url does not match any above route => not found error
