+++ src/catalog/user/user.controller.js
@@ -6,11 +6,14 @@ import { MESSAGE } from '../../constants'
 
 const NAMESPACE = `userController-${moment.utc().toISOString()}`
 
-export const findUserById = async (req, res, next) => {
+export const findUserByIdForPartner = async (req, res, next) => {
   try {
     const { id } = req.params
 
     const userInstance = await userService.findUserById(id)
+    delete userInstance.isDeleted
+    delete userInstance.createdAt
+    delete userInstance.updatedAt
 
     return res.status(httpStatusCodes.OK).json({
       message: MESSAGE.OK,
